perf(ws): decode TradeMessage.fee_rate_bps as u32 at deserialize time
Before: `pub fee_rate_bps: Option<Decimal>` required consumers (e.g. our
cryptobot WS bridge at crates/cryptobot/src/main.rs:722-724) to do a
`.to_string().parse::<u32>()` round-trip on every trade event.

After: `#[serde_as(as = "Option<DisplayFromStr>")] pub fee_rate_bps:
Option<u32>`. Fees are always integer basis points in [0, 10_000] so
Decimal precision was unnecessary overhead on the fast WS trade path.

This is a minor breaking API change for anyone consuming fee_rate_bps
as Decimal — trivial call-site fix (`.unwrap_or(Decimal::ZERO)` →
`.unwrap_or(0)`). The cryptobot call site is updated in the downstream
workspace commit that bumps the fork rev.

The matching `LastTradePrice` struct (line 171) keeps Option<Decimal>
intentionally — that path is not hot and the existing test shape uses
Decimal::ZERO assertions.

Tests: 96 lib tests pass (including all WS parse fixtures), 12 V2
signing byte-parity tests pass.

+++ src/clob/ws/types/response.rs
@@ -374,9 +374,15 @@ pub struct TradeMessage {
     /// Array of maker order details
     #[serde(default)]
     pub maker_orders: Vec<MakerOrder>,
-    /// Fee rate in basis points (string in API response)
+    /// Fee rate in basis points (serialized as a JSON string on the wire).
+    ///
+    /// Was `Option<Decimal>` in earlier versions; fees are always integer
+    /// basis points in [0, 10_000], so decoding directly to `u32` avoids a
+    /// Decimal -> String -> u32 round-trip in hot-path WS consumers (see
+    /// cryptobot's WS bridge).
+    #[serde_as(as = "Option<DisplayFromStr>")]
     #[serde(default)]
-    pub fee_rate_bps: Option<Decimal>,
+    pub fee_rate_bps: Option<u32>,
     /// On-chain transaction hash
     #[serde_as(as = "NoneAsEmptyString")]
     #[serde(default)]
